# backend/openapi.yml
openapi: 3.0.3
info:
  title: Kickstat API
  description: Sport Coaching Tool API for managing squads, athletes, events, live match logging, leagues, tactics, drills, and public squad pages.
  version: 1.1.0
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://kickstat-api-i2rc.onrender.com
    description: Production
paths:
...
  /api/reports/season:
    get:
      summary: Season report data (record, results, players, opponents)
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          description: "{ squadName, from, to, summary, matches, opponents, players }"
        '400':
          description: Malformed or reversed dates
  /api/reports/match/{kind}/{id}:
    get:
      summary: Match report data for one regular match or league fixture
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string, enum: [event, fixture] } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: "{ squadName, match, score, result, stats, scorers, cards, penalties, substitutions, timeline }"
        '404':
          description: Not a match your squad played
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Clerk JWT token
...